docs: work in progress
