+++ packages/react-docs/pages/components/transitions/slide.mdx
@@ -52,14 +52,14 @@ function Example() {
           };
 
           return (
-            <SelectableButton
+            <SelectButton
               key={value}
-              selected={value === direction}
+              isSelected={value === direction}
               onClick={onClick}
               minWidth="15x"
             >
               {value}
-            </SelectableButton>
+            </SelectButton>
           );
         })}
       </ButtonGroup>
